Moved code into new file and customized SysApiApplication
